backend/requirements.txt
fastapi==0.115.1
sqlalchemy==2.0.38
uvicorn==0.34.0
pydantic-settings==2.8.1
psycopg2-binary==2.9.10
alembic==1.15.1
requests==2.32.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
pydantic[email]==2.8.1
python-multipart==0.0.6
